Phase B7: extract diagnose handler into src/commands/diagnose.rs
