+++ crates/defra-agent-cli/src/commands/mod.rs
@@ -1,6 +1,7 @@
 // Subcommand handler implementations. Populated by Phases B4-B7.
 pub(crate) mod chat;
 pub(crate) mod config;
+pub(crate) mod diagnose;
 pub(crate) mod init;
 pub(crate) mod p2p;
 pub(crate) mod request;
